trocando os valores de Double para BigDecimal

+++ src/main/java/com/FelipeMovio/banco/dto/TransacaoDto.java
@@ -5,7 +5,7 @@ import jakarta.validation.constraints.NotNull;
 import java.math.BigDecimal;
 
 public record TransacaoDto(
-        @NotNull Double value,
+        @NotNull BigDecimal value,
         @NotNull Long payer,
         @NotNull Long payee
 ) {
